refactor(pagination): refactored pagination to namespace pattern
