+++ lib/src/Pagination/NumberedPagination.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { NumberedPagination } from ".";
-import { act, Simulate } from "react-dom/test-utils";
+import { act } from "react-dom/test-utils";
 import { unmountComponentAtNode, render } from "react-dom";
 
 describe("Component: Pagination page", () => {
