+++ lib/src/Pagination/Pagination.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import classnames from "classnames";
-import { Page, PageProps } from ".";
+import { Page, PageProps } from "./Page";
 import "./pagination.scss";
 
 const ChevronLeftIcon: JSX.Element = (
@@ -41,115 +41,112 @@ export type PaginationProps = JSX.IntrinsicElements["nav"] & {
     value?: number;
 };
 
-export const Pagination: React.FunctionComponent<PaginationProps> = React.memo(
-    ({ navs = {}, offset = 5, onPageChange, size = "md", useDotNav, showFirstAndLast: useFirstAndLast, value = 0, ...props }: PaginationProps) => {
-        const total: number = React.Children.count(props.children);
-        const indexOfLastItem: number = total - 1;
-        const disablePrev: boolean = total < 2 || value === 0;
-        const disableNext: boolean = total < 2 || value === indexOfLastItem;
+const Pagination = ({ navs = {}, offset = 5, onPageChange, size = "md", useDotNav, showFirstAndLast: useFirstAndLast, value = 0, ...props }: PaginationProps) => {
+    const total: number = React.Children.count(props.children);
+    const indexOfLastItem: number = total - 1;
+    const disablePrev: boolean = total < 2 || value === 0;
+    const disableNext: boolean = total < 2 || value === indexOfLastItem;
 
-        const renderPages = (): React.ReactElement[] => {
-            const childrenArray: React.ReactElement[] =
-                React.Children.map(props.children, (Child: React.ReactElement<PageProps>, i: number) =>
-                    React.isValidElement<PageProps>(Child)
-                        ? React.cloneElement<any>(Child, {
-                              "data-active": value === i,
-                              "data-index-number": i,
-                              key: i,
-                              onClick: (e: React.MouseEvent<HTMLLIElement>) => {
-                                  onPageChange && onPageChange(parseInt(e.currentTarget.dataset.indexNumber, 10));
-                              },
-                          })
-                        : Child
-                ) || [];
+    const renderPages = (): React.ReactElement[] => {
+        const childrenArray: React.ReactElement[] =
+            React.Children.map(props.children, (Child: React.ReactElement<PageProps>, i: number) =>
+                React.isValidElement<PageProps>(Child)
+                    ? React.cloneElement<any>(Child, {
+                          "data-active": value === i,
+                          "data-index-number": i,
+                          key: i,
+                          onClick: (e: React.MouseEvent<HTMLLIElement>) => {
+                              onPageChange && onPageChange(parseInt(e.currentTarget.dataset.indexNumber, 10));
+                          },
+                      })
+                    : Child
+            ) || [];
 
-            if (offset) {
-                /** The distance between the current value and the offset from the left. Example: ...👉|3|4|👈|(5)|6|7|... */
-                const offsetToValue: number = value - Math.floor(offset / 2);
-                /** The distance between the current value and the offset from the right. Example: ...|3|4|(5)|👉|6|7|👈... */
-                const valueToOffset: number = value + Math.floor(offset / 2);
+        if (offset) {
+            /** The distance between the current value and the offset from the left. Example: ...👉|3|4|👈|(5)|6|7|... */
+            const offsetToValue: number = value - Math.floor(offset / 2);
+            /** The distance between the current value and the offset from the right. Example: ...|3|4|(5)|👉|6|7|👈... */
+            const valueToOffset: number = value + Math.floor(offset / 2);
 
-                let offsetFrom: number = offsetToValue;
-                let offsetTo: number = valueToOffset;
+            let offsetFrom: number = offsetToValue;
+            let offsetTo: number = valueToOffset;
 
-                if (offsetToValue < 0) {
-                    offsetTo += 0 - offsetToValue;
-                    offsetTo = offsetTo > indexOfLastItem ? indexOfLastItem : offsetTo;
-                }
-                if (valueToOffset > indexOfLastItem) {
-                    offsetFrom -= valueToOffset - indexOfLastItem;
-                    offsetFrom = offsetFrom < 0 ? 0 : offsetFrom;
-                }
-
-                let filteredArray: React.ReactElement[] = childrenArray.filter((_: any, i: number) => i >= offsetFrom && i <= offsetTo);
-
-                if (!useDotNav) {
-                    if (parseInt(filteredArray[0]?.props["data-index-number"], 10) > 0) {
-                        filteredArray = [
-                            <Page className="pre-ellipsis" key="pre-ellipsis" data-disabled href={props.children[0]?.props?.href}>
-                                ...
-                            </Page>,
-                            ...filteredArray,
-                        ];
-                    }
-                    if (parseInt(filteredArray[filteredArray.length - 1]?.props["data-index-number"], 10) < indexOfLastItem) {
-                        filteredArray.push(
-                            <Page className="post-ellipsis" key="post-ellipsis" data-index-number={indexOfLastItem} data-disabled>
-                                ...
-                            </Page>
-                        );
-                    }
-                }
-                return filteredArray;
-            } else {
-                return childrenArray;
+            if (offsetToValue < 0) {
+                offsetTo += 0 - offsetToValue;
+                offsetTo = offsetTo > indexOfLastItem ? indexOfLastItem : offsetTo;
             }
-        };
+            if (valueToOffset > indexOfLastItem) {
+                offsetFrom -= valueToOffset - indexOfLastItem;
+                offsetFrom = offsetFrom < 0 ? 0 : offsetFrom;
+            }
 
-        const filteredPages = renderPages();
+            let filteredArray: React.ReactElement[] = childrenArray.filter((_: any, i: number) => i >= offsetFrom && i <= offsetTo);
 
-        const showFirst: boolean = (useFirstAndLast && !useDotNav) || (useDotNav && !disablePrev && filteredPages[0].props["data-index-number"] != 0);
-        const showLast: boolean = (useFirstAndLast && !useDotNav) || (useDotNav && !disableNext && filteredPages[filteredPages.length - 1].props["data-index-number"] != indexOfLastItem);
-        const disableFirst: boolean = disablePrev || filteredPages[0].key !== "pre-ellipsis";
-        const disableLast: boolean = disableNext || filteredPages[filteredPages.length - 1].key !== "post-ellipsis";
+            if (!useDotNav) {
+                if (parseInt(filteredArray[0]?.props["data-index-number"], 10) > 0) {
+                    filteredArray = [
+                        <Page className="pre-ellipsis" key="pre-ellipsis" data-disabled href={props.children[0]?.props?.href}>
+                            ...
+                        </Page>,
+                        ...filteredArray,
+                    ];
+                }
+                if (parseInt(filteredArray[filteredArray.length - 1]?.props["data-index-number"], 10) < indexOfLastItem) {
+                    filteredArray.push(
+                        <Page className="post-ellipsis" key="post-ellipsis" data-index-number={indexOfLastItem} data-disabled>
+                            ...
+                        </Page>
+                    );
+                }
+            }
+            return filteredArray;
+        } else {
+            return childrenArray;
+        }
+    };
 
-        return (
-            <nav {...props} className={classnames("rc", props.className)}>
-                <ul className={classnames("pagination", { [`pagination-${size}`]: size, dotnav: useDotNav })}>
-                    {props.children && (
-                        <>
-                            {showFirst && (
-                                <Page className="first-nav" onClick={() => !disableFirst && onPageChange(0)} data-disabled={disableFirst} href={props.children[0]?.props?.href}>
-                                    {navs?.first || ChevronDoubleLeftIcon}
-                                </Page>
-                            )}
-                            {!useDotNav && (
-                                <Page className="previous-nav" onClick={() => !disablePrev && onPageChange(value - 1)} data-disabled={disablePrev} href={props.children[value - 1]?.props?.href}>
-                                    {navs?.previous || ChevronLeftIcon}
-                                </Page>
-                            )}
+    const filteredPages = renderPages();
 
-                            {filteredPages}
+    const showFirst: boolean = (useFirstAndLast && !useDotNav) || (useDotNav && !disablePrev && filteredPages[0].props["data-index-number"] != 0);
+    const showLast: boolean = (useFirstAndLast && !useDotNav) || (useDotNav && !disableNext && filteredPages[filteredPages.length - 1].props["data-index-number"] != indexOfLastItem);
+    const disableFirst: boolean = disablePrev || filteredPages[0].key !== "pre-ellipsis";
+    const disableLast: boolean = disableNext || filteredPages[filteredPages.length - 1].key !== "post-ellipsis";
 
-                            {!useDotNav && (
-                                <Page className="next-nav" onClick={() => !disableNext && onPageChange(value + 1)} data-disabled={disableNext} href={props.children[value + 1]?.props?.href}>
-                                    {navs?.next || React.cloneElement(ChevronLeftIcon, { className: "h-flipped" })}
-                                </Page>
-                            )}
-                            {showLast && (
-                                <Page
-                                    className="last-nav"
-                                    onClick={() => !disableLast && onPageChange(indexOfLastItem)}
-                                    data-disabled={disableLast}
-                                    href={props.children[indexOfLastItem]?.props?.href}
-                                >
-                                    {navs?.last || React.cloneElement(ChevronDoubleLeftIcon, { className: "h-flipped" })}
-                                </Page>
-                            )}
-                        </>
-                    )}
-                </ul>
-            </nav>
-        );
-    }
-);
+    return (
+        <nav {...props} className={classnames("rc", props.className)}>
+            <ul className={classnames("pagination", { [`pagination-${size}`]: size, dotnav: useDotNav })}>
+                {props.children && (
+                    <>
+                        {showFirst && (
+                            <Page className="first-nav" onClick={() => !disableFirst && onPageChange(0)} data-disabled={disableFirst} href={props.children[0]?.props?.href}>
+                                {navs?.first || ChevronDoubleLeftIcon}
+                            </Page>
+                        )}
+                        {!useDotNav && (
+                            <Page className="previous-nav" onClick={() => !disablePrev && onPageChange(value - 1)} data-disabled={disablePrev} href={props.children[value - 1]?.props?.href}>
+                                {navs?.previous || ChevronLeftIcon}
+                            </Page>
+                        )}
+
+                        {filteredPages}
+
+                        {!useDotNav && (
+                            <Page className="next-nav" onClick={() => !disableNext && onPageChange(value + 1)} data-disabled={disableNext} href={props.children[value + 1]?.props?.href}>
+                                {navs?.next || React.cloneElement(ChevronLeftIcon, { className: "h-flipped" })}
+                            </Page>
+                        )}
+                        {showLast && (
+                            <Page className="last-nav" onClick={() => !disableLast && onPageChange(indexOfLastItem)} data-disabled={disableLast} href={props.children[indexOfLastItem]?.props?.href}>
+                                {navs?.last || React.cloneElement(ChevronDoubleLeftIcon, { className: "h-flipped" })}
+                            </Page>
+                        )}
+                    </>
+                )}
+            </ul>
+        </nav>
+    );
+};
+
+Pagination.Item = Page;
+
+export { Pagination };
