+++ lib/src/Pagination/index.ts
@@ -1,3 +1,3 @@
 export * from "./Pagination";
-export * from "./Page";
+export { PageProps } from "./Page";
 export * from "./NumberedPagination";
